pyproject.toml
[project]
name = "docgen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.20",
    "clang>=20.1.5",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.27",
    "libclang>=18.1.1",
    "numpy>=2.3.2",
    "ollama>=0.5.3",
    "pandas>=2.3.2",
    "pymupdf>=1.26.4",
    "pyyaml>=6.0.2",
    "sentence-transformers>=5.1.0",
]
